src/utils/config.py
"""
Configuration loading utilities.

This module provides functions for loading and
managing configuration files.
"""

import yaml
from pathlib import Path
from typing import Dict, Any


def load_config(config_path: str) -> Dict[str, Any]:
    """
    Load configuration from YAML file.
    
    Args:
        config_path: Path to configuration file
        
    Returns:
        Configuration dictionary
    """
    # TODO: Implement config loading
    raise NotImplementedError("load_config not implemented yet")


def merge_configs(base_config: Dict, override_config: Dict) -> Dict[str, Any]:
    """
    Merge two configuration dictionaries.
    
    Override config takes precedence over base config.
    
    Args:
        base_config: Base configuration
        override_config: Override configuration
        
    Returns:
        Merged configuration
    """
    # TODO: Implement config merging
    raise NotImplementedError("merge_configs not implemented yet")


def save_config(config: Dict[str, Any], save_path: str):
    """
    Save configuration to YAML file.
    
    Args:
        config: Configuration dictionary
        save_path: Path to save file
    """
    # TODO: Implement config saving
    raise NotImplementedError("save_config not implemented yet")
