requirements.txt
# Core dependencies
numpy>=1.21.0
scipy>=1.7.0
matplotlib>=3.5.0

# PyTorch and PyTorch Geometric
torch>=2.0.0
torch-geometric>=2.3.0

# Configuration
pyyaml>=6.0
omegaconf>=2.3.0

# Logging and visualization
tensorboard>=2.12.0
tqdm>=4.65.0

# Jupyter notebooks
jupyter>=1.0.0
ipykernel>=6.0.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
isort>=5.12.0

# Optional: Animation
# ffmpeg-python>=0.2.0
